# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LUXIMA API Documentation",
        "description": "This is the API documentation for the Luxima project. It covers all endpoints related to authentication and other functionalities.",
        "contact": {
            "email": "admin@luxima.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.luxima.id/api",
            "description": "LUXIMA API Server"
        }
    ],
    "paths": {
...
        "/category/{slug}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show detail category by slug",
                "description": "Retrieve detail data category by slug.",
                "operationId": "22ac95e86319ac642d72e9ed0fe15a56",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Unique category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Category Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/location/{slug}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Show detail location by slug",
                "description": "Retrieve detail data location by slug.",
                "operationId": "c9d5f519cca98fa219c78b5e32f099dd",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Unique location slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Vendor Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
        "/vendors": {
            "get": {
                "tags": [
                    "Vendor"
                ],
                "summary": "Show all vendors",
                "description": "Retrieve a list of all vendors.",
                "operationId": "1d0aa57c9db3be1d9403f33a7c041af2",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved vendors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vendor"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, token tidak valid"
                    }
                }
            }
        },
        "/search": {
            "get": {
                "tags": [
                    "Vendor"
                ],
                "summary": "Search for vendors based on keyword, category, and location",
                "description": "Search for vendors based on a keyword, category, and location.",
                "operationId": "bb50fd6bf54b7f73e6c0dcee3a58f54b",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Slug by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Lokasi vendor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Result Search",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vendor"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    }
                }
            }
        },
        "/vendor/{slug}": {
            "get": {
                "tags": [
                    "Vendor"
                ],
                "summary": "Show detail vendor by slug",
                "description": "Retrieve detail data vendor by slug.",
                "operationId": "3b0134ad8ea02ab1464466cac22c46bc",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Unique vendor slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Vendor Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vendor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vendor not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Categories",
                "required": [
                    "name",
                    "slug",
                    "description",
                    "image"
                ],
                "properties": {
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Unique slug for category",
                        "type": "string"
                    },
                    "description": {
                        "description": "Category description",
                        "type": "string"
                    },
                    "image": {
                        "description": "Category image",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Location": {
                "title": "Location",
                "required": [
                    "id",
                    "name",
                    "slug",
                    "location"
                ],
                "properties": {
                    "id": {
                        "description": " ID vendor",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Location name",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Unique slug for vendor",
                        "type": "string"
                    },
                    "description": {
                        "description": "Location description",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location location",
                        "type": "string"
                    },
                    "category": {
                        "description": "Location category",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Location created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Location updated at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Location deleted at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Vendor": {
                "title": "Vendor",
                "required": [
                    "id",
                    "name",
                    "slug",
                    "location"
                ],
                "properties": {
                    "id": {
                        "description": " ID vendor",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Vendor name",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Unique slug for vendor",
                        "type": "string"
                    },
                    "description": {
                        "description": "Vendor description",
                        "type": "string"
                    },
                    "location": {
                        "description": "Vendor location",
                        "type": "string"
                    },
                    "category": {
                        "description": "Vendor category",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Vendor created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Vendor updated at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Vendor deleted at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}